rawspeed: merge in warnings fixes from klaus

+++ src/external/rawspeed/RawSpeed/ErfDecoder.cpp
@@ -85,9 +85,9 @@ void ErfDecoder::decodeMetaDataInternal(CameraMetaData *meta) {
     if (wb->count == 256) {
       const ushort16 *tmp = wb->getShortArray()+24;
       // Magic values taken directly from dcraw
-      mRaw->metadata.wbCoeffs[0] = (float) tmp[0] * 508 * 1.078 / 0x10000;
+      mRaw->metadata.wbCoeffs[0] = (float) tmp[0] * 508.0f * 1.078f / (float)0x10000;
       mRaw->metadata.wbCoeffs[1] = 1.0f;
-      mRaw->metadata.wbCoeffs[2] = (float) tmp[1] * 382 * 1.173 / 0x10000;
+      mRaw->metadata.wbCoeffs[2] = (float) tmp[1] * 382.0f * 1.173f / (float)0x10000;
     }
   }
 }
